backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-dotenv
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
psycopg2-binary
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
anthropic
pdfplumber
PyPDF2
python-docx
pandas
puter-python-sdk>=0.0.18
email-validator